Fix some unsafe narrowing warnings, primarily with tables.
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3747422
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1021521}

+++ ui/views/controls/table/table_grouper.h
@@ -11,8 +11,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace views {
 
 struct VIEWS_EXPORT GroupRange {
-  int start;
-  int length;
+  size_t start;
+  size_t length;
 };
 
 // TableGrouper is used by TableView to group a set of rows and treat them
@@ -20,7 +20,7 @@ struct VIEWS_EXPORT GroupRange {
 // together.
 class VIEWS_EXPORT TableGrouper {
  public:
-  virtual void GetGroupRange(int model_index, GroupRange* range) = 0;
+  virtual void GetGroupRange(size_t model_index, GroupRange* range) = 0;
 
  protected:
   virtual ~TableGrouper() = default;
